// client/src/features/members/member-profile/member-profile.ts
import { Component, HostListener, inject, OnDestroy, OnInit, signal, ViewChild, viewChild } from '@angular/core';
import { ActivatedRoute } from '@angular/router';
import { EditableMember, Member } from '../../../types/member';
import { DatePipe } from '@angular/common';
import { MemberService } from '../../../core/services/member-service';
import { FormsModule, NgForm } from '@angular/forms';
import { ToastService } from '../../../core/services/toast-service';

@Component({
  selector: 'app-member-profile',
  imports: [DatePipe,FormsModule],
  templateUrl: './member-profile.html',
  styleUrl: './member-profile.css',
})
export class MemberProfile implements OnInit, OnDestroy {

  @ViewChild('editForm') editForm?:NgForm;
  @HostListener('window:beforeunload',['$event']) notify($event:BeforeUnloadEvent){
    if(this.editForm?.dirty){
      $event.preventDefault();
    }
  }
  private route = inject(ActivatedRoute);
  private toast = inject(ToastService);
  protected member = signal<Member | undefined>(undefined);
  protected memberService = inject(MemberService);
  protected editableMember:EditableMember={
    displayName:'',
    description:'',
    city:'',
    country:''
  };

  constructor(){
    
  }

  ngOnInit(): void {
    console.log('Member data in profile:');
    this.route.parent?.data.subscribe(data => {
      this.member.set(data['member']);
      console.log('Member data in profile:', this.member());
    })

    this.editableMember = {
      displayName:this.member()?.displayName || '',
      description:this.member()?.description || '',
      city:this.member()?.city || '',
      country:this.member()?.country || '',
    }
  }
  ngOnDestroy(): void {
    if(this.memberService.editMode()){
      this.memberService.editMode.set(false);
    }
  }
  updateProfile(){
    if(!this.member()) return;
    const updateProfile = {...this.member(), ...this.editableMember}
    console.log(updateProfile);
    this.toast.success('Profile updated successfully');
    this.memberService.editMode.set(false);
  }
}
